Crate timeq

Structs

QuadWheelWithOverflow

An implementation of four-level byte-sized wheel

TimeQ

A TimeQ is a queue datastructure where the contained items are time ordered. The underlying storage is a hashed hierarchical timer wheel, which allows for relatively cheap insertion and popping of ready items. It is also possible to cancel an entry given its id.

Enums

PopResult

TimerError

Errors encounted by a timer implementation

Traits

TimerEntryWithDelay

A trait for timer entries that store their delay along the with the state