Struct timeq::TimeQ

pub struct TimeQ<EntryType: CancellableTimerEntry + TimerEntryWithDelay> { /* private fields */ }

A TimeQ is a queue datastructure where the contained items are time ordered. The underlying storage is a hashed hierarchical timer wheel, which allows for relatively cheap insertion and popping of ready items. It is also possible to cancel an entry given its id.

Implementations

impl<EntryType: CancellableTimerEntry + TimerEntryWithDelay> TimeQ<EntryType>

pub fn new() -> Self

pub fn is_empty(&self) -> bool

Returns true if the wheel is empty

pub fn len(&self) -> usize

pub fn insert( &mut self, entry: Arc<EntryType> ) -> Result<(), TimerError<Arc<EntryType>>>

Insert a new entry

pub fn cancel_by_id( &mut self, id: &EntryType::Id ) -> Result<(), TimerError<Infallible>>

Cancel an item given its id

pub fn pop(&mut self) -> PopResult<EntryType>

Returns the set of items that need immediate action

pub fn drain(&mut self) -> Vec<Arc<EntryType>>

Drains the entire contents of the queue, returning all of the contained items