timeq

Trait TimerEntryWithDelay

pub trait TimerEntryWithDelay: Debug {
    // Required method
    fn delay(&self) -> Duration;
}

A trait for timer entries that store their delay along the with the state

Required Methods

fn delay(&self) -> Duration

Returns the time until the timeout is supposed to be triggered